refactor(gateway): extract stream-rendering + prompt-formatting mixins from base.py (shard s3)
